Working in both android and linux

+++ patterns/cnoise.glsl
@@ -1,7 +1,7 @@
-#version 330 core
+#version 310 es
 
 #ifdef GL_ES
-precision mediump float;
+precision highp float;
 #endif
 #define PI 3.1415926535897932384626433832795
